Replace bios with shorter text and add Fellows' start date info
Fellows and Innovation staff were asked to rewrite their bios with a
character-limited “Reason for Joining” statement to create a more
consistent layout. Start dates were added to all of the Fellows to
assist in ordering on the Home page and on the About page. The Home
page is ordered by most recent Fellow first. The About page is ordered
by longest-serving Fellow first.